refactor hero block
