+++ src/blocks/hero/edit.js
@@ -6,11 +6,11 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import { title } from '../';
+import { title } from './';
 import Inspector from './inspector';
 import Controls from './controls';
 import applyWithColors from './colors';
-import { BackgroundStyles, BackgroundClasses, BackgroundVideo, BackgroundDropZone } from '../../../components/background';
+import { BackgroundStyles, BackgroundClasses, BackgroundVideo, BackgroundDropZone } from '../../components/background';
 
 /**
  * WordPress dependencies
