+++ src/blocks/hero/icons.js
