+++ src/blocks/hero/index.js
@@ -1,16 +1,13 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * Internal dependencies
  */
 import './styles/style.scss';
 import './styles/editor.scss';
-import icons from './components/icons';
-import Edit from './components/edit';
-import { BackgroundAttributes, BackgroundClasses, BackgroundVideo } from '../../components/background';
+import icons from './icons';
+import edit from './edit';
+import save from './save';
+import transforms from './transforms';
+import { BackgroundAttributes } from '../../components/background';
 import DimensionsAttributes from '../../components/dimensions-control/attributes';
 import CSSGridAttributes from '../../components/grid-control/attributes';
 import ResponsiveBaseControlAttributes from '../../components/responsive-base-control/attributes';
@@ -19,8 +16,6 @@ import ResponsiveBaseControlAttributes from '../../components/responsive-base-co
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { createBlock } = wp.blocks;
-const { getColorClassName, InnerBlocks } = wp.blockEditor;
 
 /**
  * Block constants
@@ -112,87 +107,11 @@ const settings = {
 		coBlocksSpacing: true,
 	},
 
-	transforms: {
-		from: [
-			{
-				type: 'prefix',
-				prefix: ':hero',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-		],
-	},
+	transforms,
 
-	edit: Edit,
+	edit,
 
-	save( { attributes } ) {
-		const {
-			coblocks,
-			layout,
-			fullscreen,
-			maxWidth,
-			backgroundImg,
-			backgroundType,
-			paddingSize,
-			backgroundColor,
-			customBackgroundColor,
-			customTextColor,
-			textColor,
-			contentAlign,
-			focalPoint,
-			hasParallax,
-			height,
-		} = attributes;
-
-		const textClass = getColorClassName( 'color', textColor );
-		const backgroundClass = getColorClassName( 'background-color', backgroundColor );
-
-		const classlist = {
-			'has-text-color': textColor || customTextColor,
-			[ textClass ]: textClass,
-			[ `coblocks-hero-${ coblocks.id }` ]: coblocks && ( typeof coblocks.id !== 'undefined' ),
-		};
-
-		const classes = classnames( classlist );
-
-		const styles = {
-			color: textClass ? undefined : customTextColor,
-		};
-
-		const innerClasses = classnames(
-			'wp-block-coblocks-hero__inner',
-			...BackgroundClasses( attributes ), {
-				[ `hero-${ layout }-align` ]: layout,
-				'has-text-color': textColor && textColor.color,
-				'has-padding': paddingSize && paddingSize !== 'no',
-				[ `has-${ paddingSize }-padding` ]: paddingSize && paddingSize !== 'advanced',
-				[ backgroundClass ]: backgroundClass,
-				[ `has-${ contentAlign }-content` ]: contentAlign,
-				'is-fullscreen': fullscreen,
-			} );
-
-		const innerStyles = {
-			backgroundColor: backgroundClass ? undefined : customBackgroundColor,
-			backgroundImage: backgroundImg && backgroundType === 'image' ? `url(${ backgroundImg })` : undefined,
-			color: textColor ? textColor.color : undefined,
-			backgroundPosition: focalPoint && ! hasParallax ? `${ focalPoint.x * 100 }% ${ focalPoint.y * 100 }%` : undefined,
-			minHeight: fullscreen ? undefined : height,
-		};
-
-		return (
-			<div className={ classes } style={ styles } >
-				<div className={ innerClasses } style={ innerStyles }>
-					{ BackgroundVideo( attributes ) }
-					<div className="wp-block-coblocks-hero__box" style={ { maxWidth: maxWidth ? maxWidth + 'px' : undefined } }>
-						<InnerBlocks.Content />
-					</div>
-				</div>
-			</div>
-		);
-	},
+	save,
 };
 
 export { name, title, icon, settings };
